Move user sub models to seperate folders
