+++ src/app/components/forms/user-profile-form/user-profile-form.component.ts
@@ -11,8 +11,8 @@ import {OnDestroy} from '@angular/core';
 import {OnInit} from '@angular/core';
 import {Subscription} from 'rxjs/Subscription';
 import {User} from '../../../models/api-models/user/user';
-import {UserDocument} from '../../../models/api-models/user/user-document';
-import {UserImage} from '../../../models/api-models/user/user-image';
+import {UserDocument} from '../../../models/api-models/user-document/user-document';
+import {UserImage} from '../../../models/api-models/user-image/user-image';
 import {UserProxy} from '../../../services/proxy/user-proxy.service';
 import {UserResolver} from '../../../resolvers/user/user.resolver';
 import {Validators} from '@angular/forms';
