+++ src/app/models/api-models/user-interest/user-interest.ts
